Add exceptional handing cases

+++ src/main/java/com/example/employeeManagement/service/EmployeeServiceImpl.java
@@ -1,6 +1,9 @@
 package com.example.employeeManagement.service;
 
+import com.example.employeeManagement.exception.DatabaseConnectionException;
+import com.example.employeeManagement.exception.DuplicateEmployeeIdException;
 import com.example.employeeManagement.exception.EmployeeNotFoundException;
+import com.example.employeeManagement.exception.InvalidIdFormatException;
 import com.example.employeeManagement.model.Employees;
 import com.example.employeeManagement.repository.EmployeeRepository;
 import org.springframework.stereotype.Service;
@@ -18,24 +21,43 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     public Employees createEmployee(Employees employees) {
+        try {
+            employeeRepository.save(employees);
+        } catch (Exception e) {
+            throw new DatabaseConnectionException("Database error while saving employee", e);
+        }
         return employeeRepository.save(employees);
     }
 
     @Override
     public Employees getEmployeeById(String id) {
+        if (id == null || id.isBlank()) {
+            throw new InvalidIdFormatException("Employee id cannot be empty");
+        }
         return employeeRepository.findById(id)
                 .orElseThrow(() -> new EmployeeNotFoundException("Employee ID not found: " + id));
     }
 
     @Override
     public List<Employees> getAllEmployees() {
-        return employeeRepository.findAll();
+        try {
+            return employeeRepository.findAll();
+        } catch (Exception e) {
+            throw new DatabaseConnectionException("unable to fetch employee", e);
+        }
     }
 
     @Override
     public Employees updateEmployee(String id, Employees updatedEmployee) {
+        if (id == null || id.isBlank()){
+            throw new InvalidIdFormatException("Employee id cannot be empty");
+        }
         Employees existing = employeeRepository.findById(id)
                 .orElseThrow(() -> new EmployeeNotFoundException("Employee ID not found: " + id));
+
+        if (!id.equals(updatedEmployee.getId()) && employeeRepository.existsById(updatedEmployee.getId())){
+            throw new DuplicateEmployeeIdException("Employee ID already exists: " + updatedEmployee.getId());
+        }
 
         existing.setFirstName(updatedEmployee.getFirstName());
         existing.setLastName(updatedEmployee.getLastName());
@@ -47,8 +69,12 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     public void deleteEmployee(String id) {
+        if (id == null || id.isBlank()) {
+            throw new InvalidIdFormatException("Employee ID cannot be empty");
+        }
+
         if (!employeeRepository.existsById(id)) {
-            throw new EmployeeNotFoundException("Employee ID not found: " + id);
+            throw new EmployeeNotFoundException("Employee not found: " + id);
         }
         employeeRepository.deleteById(id);
     }
